Move draw_data to utilities and make public.
