+++ src/thesis/simple_model/funcs.py
@@ -11,6 +11,7 @@ from statsmodels.api import add_constant  # type: ignore[import-untyped]
 from statsmodels.sandbox.regression.gmm import IV2SLS  # type: ignore[import-untyped]
 
 from thesis.classes import Instrument, LocalATEs
+from thesis.utilities import draw_data
 
 
 def simulation_bootstrap(
@@ -61,7 +62,7 @@ def simulation_bootstrap(
     beta_hi = np.zeros(n_sims)
 
     for i in range(n_sims):
-        data = _draw_data(n_obs, local_ates=local_ates, instrument=instrument, rng=rng)
+        data = draw_data(n_obs, local_ates=local_ates, instrument=instrument, rng=rng)
 
         beta_lo[i], beta_hi[i] = _idset(
             b_late=_late(data),
@@ -406,35 +407,6 @@ def _late_2sls(data: np.ndarray) -> tuple[float, float]:
     res = model.fit()
 
     return res.params[1], res.bse[1]
-
-
-def _draw_data(
-    n_obs,
-    local_ates: LocalATEs,
-    instrument: Instrument,
-    rng: np.random.Generator,
-) -> np.ndarray:
-    z = rng.choice(instrument.support, size=n_obs, p=instrument.pmf)
-
-    u = rng.uniform(low=0, high=1, size=n_obs)
-
-    d = np.where(z == 1, u <= instrument.pscores[1], u <= instrument.pscores[0])
-
-    _never_takers = u <= instrument.pscores[0]
-    _compliers = (u > instrument.pscores[0]) & (u <= instrument.pscores[1])
-    _always_takers = u > instrument.pscores[1]
-
-    y0 = np.zeros(n_obs)
-
-    y1 = (
-        _never_takers * local_ates.never_taker
-        + _compliers * local_ates.complier
-        + _always_takers * local_ates.always_taker
-    )
-
-    y = d * y1 + (1 - d) * y0 + rng.normal(scale=0.1, size=n_obs)
-
-    return np.column_stack((y, d, z, u))
 
 
 def _draw_bootstrap_data(
